http://luc.lino-framework.org/blog/2018/0105.html

+++ lino_xl/lib/tim2lino/spzloader2.py
@@ -209,7 +209,7 @@ class TimLoader(TimLoader):
         except Course.DoesNotExist:
             dd.logger.warning("Unknown PLP->IdPar %s", idpar)
             return
-        kw.update(partner=course.partner)
+        kw.update(partner=course.partner.person)
         yield Guest(**kw)
                 
 
@@ -302,11 +302,13 @@ class TimLoader(TimLoader):
 
         # yield self.load_dbf('PLP')
 
+        if False:  # temporarily deactivaed
+            Event = rt.models.cal.Event
+            Event.objects.all().delete()
+            yield self.load_dbf('DLS')
+        
         Guest = rt.models.cal.Guest
         Guest.objects.all().delete()
-        Event = rt.models.cal.Event
-        Event.objects.all().delete()
-        yield self.load_dbf('DLS')
         yield self.load_dbf('DLP')
         yield self.load_dbf('PAR')
